reducing noise
