+++ packages/strapi-design-system/src/KeyboardNavigable/index.js
@@ -1,0 +1,1 @@
+export * from './KeyboardNavigable';
